Cleanup: move core models into package
More models are meant to live here so let's make room for them.
