+++ pootle/core/models/__init__.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright (C) Pootle contributors.
+#
+# This file is a part of the Pootle project. It is distributed under the GPL3
+# or later license. See the LICENSE file for a copy of the license and the
+# AUTHORS file for copyright and authorship information.
+
+from .revision import Revision
+from .virtualresource import VirtualResource
+
+
+__all__ = ('Revision', 'VirtualResource')
